# docs/openapi-banco-orbital.yaml
...
openapi: 3.0.3
info:
  title: Banco Orbital — API
  version: 1.0.0-draft
  description: |
    API del home banking de **Banco Orbital**. Este documento es el contrato entre
    nuestro backend y nuestro frontend, y la referencia para los otros bancos del
    curso que quieran comparar.

    ## Alcance de la v1

    Cuentas, tarjetas, préstamos y plazos fijos. **Seguros, CEDEARs y reservas
    quedan para la v2** — decisión tomada para que el alcance entre en el tiempo
    disponible.

    Los proveedores externos simulados (recargas de celular y empresas de
    servicios) **no están acá**: viven en una API aparte, `banco-proveedores`, que
    este banco consume por HTTP como si fuera un tercero real.

    ## Convenciones

    | Tema | Convención |
    |---|---|
    | Autenticación | JWT de Clerk en `Authorization: Bearer <token>` |
    | Errores | `{ "error": "mensaje" }`, con `details` opcional. Igual que el Banco Central |
    | Paginación | `page` y `limit` por query; respuesta `{ page, limit, count, data }` |
    | Importes | `number` con 2 decimales. Internamente se calculan en decimal exacto |
    | Nombres | Español, según `docs/GLOSARIO.md`. Campos en `snake_case`, como ya los devuelven los DTOs |
    | Idempotencia | Header `Idempotency-Key` (UUID v4) en **todo endpoint que mueve plata** |

    ### Sobre `monto` contra `importe`

    Dentro de nuestro banco el campo siempre se llama **`monto`**. `importe` aparece
...
    contrato. No mezclarlos.

    ### Sobre las rutas
...
servers:
...
security:
  - ClerkJWT: []
...
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de sesión emitido por Clerk. El backend lo valida y resuelve el
        `usuario` y su `persona` asociada. Todos los endpoints bajo `/api` lo
        requieren, además de usuario activo y perfil completo.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
      description: Página, empezando en 1.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Cantidad de elementos por página.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        UUID v4. Protege contra el doble envío: reintentar con la misma clave
        devuelve la respuesta original en vez de ejecutar la operación otra vez.
      example: 3f2504e0-4f89-41d3-9a0c-0305e82c3301
  responses:
    NoAutorizado:
      description: Falta el token, es inválido, o el usuario está inactivo.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No autorizado.' }
    Prohibido:
      description: El usuario no tiene permiso, o el recurso pertenece a otra persona.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No puedes operar sobre una cuenta que no te pertenece.' }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'No existe el recurso solicitado.' }
    ErrorValidacion:
      description: El cuerpo o los parámetros no pasaron la validación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: 'Datos inválidos.'
            details:
              fieldErrors:
                monto: ['El monto debe ser mayor a cero.']
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensaje legible. En producción los 5xx se reemplazan por uno genérico.
          example: 'Saldo insuficiente.'
        details:
          type: object
          nullable: true
          description: |
            Detalle adicional. Presente en errores de validación (con `fieldErrors`)
            y en desarrollo. Nunca expone internos de la base en producción.
      additionalProperties: false
    Paginado:
      type: object
      description: Envoltorio de todos los listados paginados.
      required: [page, limit, count, data]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        count:
          type: integer
          description: Cantidad de elementos en **esta** página, no el total del recurso.
          example: 20
        data:
          type: array
          items: { type: object }
    Moneda:
      type: string
      enum: [ARS, USD]
      description: Moneda de la cuenta. La caja en pesos nace con la persona; la de dólares se abre aparte.
      example: ARS
    Monto:
      type: number
      format: double
      description: |
        Importe con 2 decimales. Se almacena como `NUMERIC(18,2)` y se calcula en
        decimal exacto; se expone como número por comodidad del cliente, que sólo
        lo formatea para mostrar.
      example: 184500.20
    Cbu:
      type: string
      minLength: 22
      maxLength: 22
      pattern: '^\d{22}$'
      description: Clave Bancaria Uniforme. Cada cuenta tiene el suyo, así que la caja en pesos y la de dólares de una misma persona tienen CBU distinto.
      example: '0060001948123456001608'
...
    CuotaPrestamo:
      type: object
      description: Fila del cronograma de amortización, sistema francés.
      properties:
        numero: { type: integer, example: 1 }
        vencimiento: { type: string, format: date, example: '2026-10-08' }
        cuota: { $ref: '#/components/schemas/Monto' }
        capital:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          description: Porción que amortiza capital.
        interes:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          description: Porción de intereses, calculada sobre el saldo remanente.
        saldo:
          allOf: [{ $ref: '#/components/schemas/Monto' }]
          description: Capital adeudado después de pagar esta cuota.
        estado:
          type: string
          enum: [pendiente, pagada, en_mora]
          description: Sólo en préstamos otorgados; ausente en las simulaciones.
    Prestamo:
      type: object
      properties:
        id: { type: string, format: uuid }
        persona_id: { type: string, format: uuid }
        cuenta_id:
          type: string
          format: uuid
          description: Cuenta donde se acreditó el capital y de donde se debitan las cuotas.
        moneda: { $ref: '#/components/schemas/Moneda' }
        capital: { $ref: '#/components/schemas/Monto' }
        cuotas: { type: integer, example: 12 }
        tna:
          type: number
          description: |
            Tasa Nominal Anual **en porcentaje** (72 significa 72 %). Ojo: ArgentinaDatos
            la publica como fracción decimal (0.72); el adapter la normaliza antes de llegar acá.
          example: 72
        tem:
          type: number
          description: Tasa Efectiva Mensual en porcentaje, `tna / 12`.
          example: 6
        cuota_mensual: { $ref: '#/components/schemas/Monto' }
        total_a_pagar: { $ref: '#/components/schemas/Monto' }
        total_intereses: { $ref: '#/components/schemas/Monto' }
        cft:
          type: number
          description: Costo Financiero Total anual en porcentaje. Sin cargos extra coincide con la TEA.
          example: 101.22
        saldo_deuda: { $ref: '#/components/schemas/Monto' }
        cuotas_pagadas: { type: integer, example: 1 }
        estado:
          type: string
          enum: [vigente, cancelado, en_mora]
          example: vigente
        fecha_otorgamiento: { type: string, format: date }
        cronograma:
          type: array
          items: { $ref: '#/components/schemas/CuotaPrestamo' }
...
paths:
...
  /api/transacciones/destinatario/resolver:
    get:
      tags: [Transacciones]
      summary: Resolver un alias o CBU antes de transferir
      description: |
        Busca el titular de un alias o CBU en el Banco Central, para mostrarle al
        cliente a quién le va a transferir antes de confirmar. Devuelve también la
        **moneda** de la cuenta, que es lo que permite validar la transferencia.
      parameters:
        - name: alias
          in: query
          schema: { type: string }
          description: Uno de los dos, alias o cbu.
        - name: cbu
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Titular encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cbu: { $ref: '#/components/schemas/Cbu' }
                  alias: { type: string, nullable: true }
                  nombre: { type: string, example: María }
                  apellido: { type: string, example: García }
                  moneda: { $ref: '#/components/schemas/Moneda' }
                  banco: { type: string, example: Banco Nodo }
        '400':
          description: No se indicó ni alias ni CBU.
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404':
          description: El alias o CBU no existe en el sistema.

  # ───────────────────────────── Tarjetas ─────────────────────────────
...
  /api/prestamos/simulaciones:
    post:
      tags: [Préstamos]
      summary: Simular un préstamo
      description: |
        Calcula la cuota y el cronograma completo **sin registrar nada**.

        **Método acordado con los otros bancos — sistema francés:**

        ```
        tem   = tna / 12 / 100
        cuota = capital × tem / (1 − (1 + tem)^−cuotas)
        ```

        - Redondeo a 2 decimales, ROUND_HALF_UP.
        - El interés de cada cuota se calcula sobre el saldo remanente.
        - La **última cuota absorbe el residuo** para que el saldo cierre exacto en 0.

        Si se omite `tna`, se usa la de referencia de ArgentinaDatos, ya normalizada
        a porcentaje.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capital, cuotas]
              properties:
                capital: { $ref: '#/components/schemas/Monto' }
                cuotas: { type: integer, minimum: 1, maximum: 72, example: 12 }
                tna: { type: number, nullable: true, example: 72 }
                moneda: { $ref: '#/components/schemas/Moneda' }
            example: { capital: 1000000, cuotas: 12, tna: 72 }
      responses:
        '200':
          description: Simulación calculada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  capital: { $ref: '#/components/schemas/Monto' }
                  cuotas: { type: integer }
                  tna: { type: number }
                  tem: { type: number }
                  cuota_mensual: { $ref: '#/components/schemas/Monto' }
                  total_a_pagar: { $ref: '#/components/schemas/Monto' }
                  total_intereses: { $ref: '#/components/schemas/Monto' }
                  cft: { type: number }
                  cronograma:
                    type: array
                    items: { $ref: '#/components/schemas/CuotaPrestamo' }
              example:
                capital: 1000000
                cuotas: 12
                tna: 72
                tem: 6
                cuota_mensual: 119277.03
                total_a_pagar: 1431324.36
                total_intereses: 431324.36
                cft: 101.22
                cronograma:
                  - { numero: 1, cuota: 119277.03, capital: 59277.03, interes: 60000, saldo: 940722.97 }
                  - { numero: 2, cuota: 119277.03, capital: 62833.65, interes: 56443.38, saldo: 877889.32 }
        '400': { $ref: '#/components/responses/ErrorValidacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
  /api/prestamos:
    post:
      tags: [Préstamos]
      summary: Solicitar un préstamo
      description: |
        Otorga el préstamo, acredita el capital en la cuenta indicada y **lo informa
        al Banco Central** con `POST /central-deudores`, para que el resto del sistema
        financiero lo vea.

        Antes de otorgar se consulta la central de deudores: **situación 3 o peor
        rechaza la solicitud**, igual que en la apertura de cuenta.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cuenta_id, capital, cuotas]
              properties:
                cuenta_id:
                  type: string
                  format: uuid
                  description: Dónde se acredita el capital. Define la moneda del préstamo.
                capital: { $ref: '#/components/schemas/Monto' }
                cuotas: { type: integer, minimum: 1, maximum: 72 }
                tna: { type: number, nullable: true }
      responses:
        '201':
          description: Préstamo otorgado y acreditado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Prestamo' }
        '400': { $ref: '#/components/responses/ErrorValidacion' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403':
          description: Sin permiso, **o** el titular está en situación 3 o peor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error: 'No se puede otorgar el préstamo: el titular registra situación 4 en la central de deudores.'
        '404': { $ref: '#/components/responses/NoEncontrado' }

    get:
      tags: [Préstamos]
      summary: Listar préstamos
      parameters:
        - name: estado
          in: query
          schema: { type: string, enum: [vigente, cancelado, en_mora] }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Préstamos visibles para el usuario. El cronograma se omite acá.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginado'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Prestamo' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
  /api/prestamos/{prestamoId}:
    get:
      tags: [Préstamos]
      summary: Detalle de un préstamo
      description: Incluye el cronograma completo con el estado de cada cuota.
      parameters:
        - name: prestamoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Préstamo encontrado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Prestamo' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/prestamos/{prestamoId}/pagos:
    post:
      tags: [Préstamos]
      summary: Pagar la próxima cuota
      description: |
        Debita la cuota de la cuenta del préstamo y la marca como pagada.

        **Reglas:** las cuotas se pagan en orden, no se aceptan pagos parciales, y
        una cuota entra en mora a los 31 días corridos de su vencimiento. Al pagar la
        última, el préstamo pasa a `cancelado` y se actualiza el informe en el Banco
        Central.
      parameters:
        - name: prestamoId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Cuota pagada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cuota: { $ref: '#/components/schemas/CuotaPrestamo' }
                  cuotas_pendientes: { type: integer, example: 11 }
                  saldo_deuda: { $ref: '#/components/schemas/Monto' }
                  estado_prestamo: { type: string, enum: [vigente, cancelado, en_mora] }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409':
          description: El préstamo ya está cancelado.
        '422':
          description: Saldo insuficiente en la cuenta.
  /api/prestamos/{prestamoId}/precancelacion:
    post:
      tags: [Préstamos]
      summary: Precancelar el préstamo
      description: |
        Cancela la deuda antes de tiempo pagando el **capital adeudado**, sin los
        intereses de las cuotas futuras. El préstamo pasa a `cancelado` y se actualiza
        el informe en el Banco Central.
      parameters:
        - name: prestamoId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Préstamo precancelado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Prestamo' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409':
          description: El préstamo ya estaba cancelado.
        '422':
          description: Saldo insuficiente para cubrir el capital adeudado.

  # ──────────────────────────── Inversiones ────────────────────────────
...
  /api/prestamos/actualizar-mora:
    post:
      tags: [Préstamos]
      summary: Barrido de mora
      description: |
        Recorre los préstamos con cuotas vencidas hace más de 31 días, los marca
        `en_mora` y **reinforma la situación al Banco Central**.

        Es la pieza que hace que la central de deudores sirva de verdad. El
        cliente que **deja** de pagar no dispara ninguna acción, así que sin este
        barrido el Banco Central seguiría informando situación 1 sobre alguien
        que hace medio año no aparece.

        Se dispara a mano o desde un cron externo. Se descartó un `setInterval`
        porque se duplica si corren dos instancias y no se puede testear.

        Sólo roles internos.
      responses:
        '200':
          description: Resumen del barrido.
          content:
            application/json:
              schema:
                type: object
                properties:
                  revisados: { type: integer, description: Préstamos con cuotas vencidas encontrados }
                  en_mora: { type: integer, description: Marcados en mora }
                  informados: { type: integer, description: Informados al Banco Central con éxito }
                  errores: { type: integer, description: Los que el Central rechazó o no respondió }
              example: { revisados: 1, en_mora: 1, informados: 1, errores: 0 }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
...